Add show_tok which summarizes the 'cost' of the current chat (until it overflows the 4000-token allowance).
Create menu system that the bot will loop back to when exiting the adventure or other adventure-builder options.

Add menu with options to play (normal), play from a previous log, generate an adventure framework, and initiate general chat with the bot.

The logfile name is now global in the event we resume an adventure and want to re-save to the same file.